feat: generic simple value tests
